[cleanup] moves Login/Register.js to containers, removes console.logs
