+++ client/containers/Singleplayer.js
@@ -123,10 +123,8 @@ class Singleplayer extends Component {
     let minutes = this.props.gameTime.minutes;
     let seconds = this.props.gameTime.seconds;
     let tenthSeconds = this.props.gameTime.tenthSeconds;
-    console.log('my minutes, seconds, tenthseconds is: ', minutes, seconds, tenthSeconds);
     let yourTime = (minutes*60 + seconds + tenthSeconds/10).toFixed(1);
     let bestTime = (highScoreObj.oldReplayDuration / 1000).toFixed(1);
-    console.log(highScoreObj);
 
     // Set title and message for sweet alert
     if (highScoreObj.newHighScore && highScoreObj.loggedIn) {
@@ -168,7 +166,6 @@ class Singleplayer extends Component {
       function(isConfirm) {
         if (isConfirm === true) {
           location.reload();
-          console.log('Confirm false, currentlevel', this.props.currentLevel);
           this.props.changeLevel({'currentLevel': null});
           this.props.changeLevel({'currentLevel': this.props.currentLevel.currentLevel});
         } else if (isConfirm === false) {
